tdf#149888 Make F1 work in Math initial state
Currently, when Math is opened and F1 is pressed in its initial state, no help page is opened. This is caused by the GrabFocus call in InitialFocusTimerHdl (see comments in its implementation).

This patch ensures that pressing F1 both in the Edit and Graphics windows will open the main help page of the Math module.

NOTE: This patch will only work after the related patch in helpcontent2 gets accepted, since it is responsible for creating the HID target in the XHP file.

Change-Id: I59980cea1816a863e1aaa2d0757d9b141437ead5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154691
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>