tdf#99609 a11y: Fix reporting selection in tab list table
The previous calculation of row and column index
in `AccessibleTabListBoxTable::getSelectedAccessibleChild`
based on the selection index was wrong, so fix that.

Only increase the the loop variable `nRow` in
`AccessibleTabListBoxTable::implGetSelRow` at the end
of the loop, since the comparison still needs to use
the original value. (Index starts at 0. Trying to retrieve
the first selected child at index 0 would otherwise
fail, and always return the fallback value of row index 0,
making NVDA always announce the first row as selected instead
of the actually selected one.

The issue could also be observed with Accerciser and the qt6
VCL plugin. With the "org.openoffice.LDAP" row selected
in the dialog, and the table object selected in Accerciser's
treeview of the a11y hierarchy, the cells in the first
row ("org.openoffice.VCL") would always be returned when
querying the selection using the AT-SPI Selection interface:

    In [1]: sel = acc.querySelection()
    In [2]: sel.nSelectedChildren
    Out[2]: 4
    In [3]: sel.getSelectedChild(0).name
    Out[3]: 'org.openoffice.VCL'

This works as expected now with this fix in place:

    In [4]: sel = acc.querySelection()
    In [5]: sel.nSelectedChildren
    Out[5]: 4
    In [6]: sel.getSelectedChild(0).name
    Out[6]: 'org.openoffice.LDAP'

With this fix in place, NVDA now properly announces the
currently focused row in the Expert Configuration
dialog (i.e. both, the focused cell first, and then
the whole row as selection).

Change-Id: I394027695616b17f7290136d8ee10c06a0282e80
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154688
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>