tdf#99609 a11y: Fix bound rect reporting for tab list box
As described in more detail for the case in
Change-Id I35f7280d2c386a9a8e04e636ebf34850a733c84a
("tdf#99609 a11y: Announce the correct entry in tab list box"),
this also needs to use `SvTabListBox::GetEntryOnPos` for the
correct entry to be used.

With this in place, Accerciser now highlights the correct entry
when clicking on the corresponding object in its treeview of
LO's a11y hierarchy.

Change-Id: I850247cd384a6a9bb909f51b39c4743548ec1105
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154689
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>