Introduce "..."_tstr as a companion to "..."_ostr/u"..."_ustr in template code
...like

>         *path = "/";

in

> template<typename T> oslFileError getSystemPathFromFileUrl(
>     OUString const & url, T * path, bool resolveHome)

in sal/osl/unx/file_url.cxx, where T is either OString or OUString.  That can
now be replaced with

>         *path = "/"_tstr;

and for OString it will be a no-cost constexpr operation, while for OUString it
will still construct the OUString instance at runtime (as did the original code
for both OString and OUString).

(This change required moving the definition of rtl::detail::OStringHolder
around in include/rtl/string.hxx.  For consistency, I similarly moved around the
definition of rtl::detail::OUStringHolder in include/rtl/ustring.hxx, too.)

Change-Id: I7fb3a81b438f98c609684f7e70b7a60068d32fdb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154748
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>