tdf#99609 a11y Set old/new value correctly in a11y event again
... after

    commit 7835f97113460922cfe14685baa7bcaea2de736a
    Date:   Tue Jul 18 16:28:08 2023 +0200

        add IndexHint -1 for 2 commitEvent methods in accessibility

had switched from using the default AccessibleEventObject ctor
and setting all of the members manually, but
presumably unintentionally swapped old and new value along
the way, since the ctor takes the new value as 3rd and the
old value as 4th argument.

I ran into this because because announcement of the items
in the Expert Configuration dialog with Orca and the qt6 VCL plugin
dialog no longer worked after merging

    commit 452c17e08fd8f18d032788170d873b019c8e9716
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Jul 21 08:27:09 2023 +0200

        tdf#99609 a11y: Announce the correct entry in tab list box

, while it did still work with the tdf#99609 commit series
on top of master just a few days old.

Change-Id: I516a3067ca04b2e4a06011148ac2690bd301fefa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154719
Reviewed-by: Julien Nabet <serval2412@yahoo.fr>
Tested-by: Jenkins