tdf#155414 include system "reduce animation" preferences
Allow the system's "reduce animation" preferences to disable the
Calc animated border when copying a selection of cells.

Also, fix failure to stop the animated border when a copy action is
performed in a different document by cancelling the animated border.
This transferable has lost ownership of the system clipboard so,
in essence, the previous copy action that created this transferable
has been cancelled.

Change-Id: I4da6d557c5dbf653c38a99eb7a16d961843a41d6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154752
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Reviewed-by: Patrick Luby <plubius@neooffice.org>