tdf#154447 Fix menu position in kf5
Prior to this patch the menus were using "topLeft" as reference for positioning the menu, which made them appear on top of the control that originated their appearance.

With this patch, it now uses "bototmLeft" to place the menu below the control (similar to gtk3).

Change-Id: I66b312fe09e078ec067773a483f9e59f627fcf93
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154771
Tested-by: Jenkins
Tested-by: Gabor Kelemen <kelemeng@ubuntu.com>
Reviewed-by: Gabor Kelemen <kelemeng@ubuntu.com>
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
(cherry picked from commit 7b060d1c2acf03378c87abb7215f9c0a957b549c)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154857
Reviewed-by: Rafael Lima <rafael.palma.lima@gmail.com>