tdf#154447 qt: Take menu bar into account for popup pos
Change-Id I66b312fe09e078ec067773a483f9e59f627fcf93
("tdf#154447 Fix menu position in kf5") fixes the
position for menu popups for the case where no menu
bar is present, e.g. the default after switching to
the Tabbed interface.

While it improves the case where a menu bar is present,
the menu would would still open a bit too far up for
that case.

Adapt the position by the menu bar height to make that
case work as expected as well.

Do this in `QtMenu::ShowNativePopupMenu` for now since that's
easiest. Potentially, the height should have been taken
into consideration somewhere earlier, but since the whole
handling of menu bar height is quite "special", don't
aim to address that elsewhere now, s.a. an earlier attempt
to rework the whole handling in

    commit afc828b9833b7a612369e95606ba56d41ef2c369
    Date:   Sat May 28 23:47:21 2022 +0200

        VCL expect correct frame size for native menubars

which was later reverted in

    commit f51b220b953ec71fb742f799fbe645a93cf3d944
    Date:   Fri Mar 24 08:06:56 2023 +0100

        tdf#149805 tdf#151677 tdf#152217 tdf#154043 tdf#153458 tdf#153800 Revert "VCL expect

    ... correct frame size for native menubars"

Change-Id: Iafd80997f9f506cb22afe1aaf6d1a3f716700ea3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154839
Tested-by: Jenkins
Tested-by: Rafael Lima <rafael.palma.lima@gmail.com>
Tested-by: Gabor Kelemen <kelemeng@ubuntu.com>
Reviewed-by: Gabor Kelemen <kelemeng@ubuntu.com>