lok: notify only affected tabs on cell resize
When row has automatic height and we type something into
one cell - it may be resized. That caused sending
document size changed callback to all the views, also
affecting all the parts/tabs.

This patch calls that only for affected views which
are looking at modified parts. It will not invalidate
all the tabs when only one is modified.

Change-Id: I93e574cf967d608696c1a95ad8b5a7614973c3f5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154891
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>