tdf#156156 Add Help button to sidebar decks
This patch adds a Help button to the sidebar deck, so that each deck has its own Help ID, which can later be used as reference to write help pages.

The Help IDs of each sidebar deck is a combination of the string "SIDEBAR_" + the deck ID (as defined in Sidebar.xcu).

Change-Id: Ib7f106ff917e41130bde136e1e6f04bb2af40daf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154918
Tested-by: Jenkins
Reviewed-by: Stéphane Guillou <stephane.guillou@libreoffice.org>
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>