tdf#129713 Output message when using "--cat" with other LO instances running
This change modifies the code that determines whether a 2nd LibreOffice
instance is running during the startup of the new soffice.bin process.
In that section of code, we simply write a new error message to console
and pop-up a message box via FatalError.  The message tells the user
that they should close other LibreOffice processes if they want to use
the "--cat" or "--script-cat" command line options.

Change-Id: I2c2a00c07d733e2f0ed6c0632f0f0d115188b116
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154909
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>