qt a11y: Return app as parent for top-level objects
If an accessible object does not have a parent in the
LO a11y hierarchy or in the QObject hierarchy, return
the app as accessible parent.
Top-level items (like dialogs) don't have another
accessible parent, but are children of the app.
This can also be seen e.g. in Accerciser: For example
Writer's "Page Style" dialog is listed as a child of the
application, but it had no parent without this change
in place, breaking automatically selecting the bookmarked
a11y object in Accerciser's treeview of the LO a11y hierarchy
when restoring the bookmark, since that relies being able to
walk up the a11y hierarchy.

This is in line with what Qt itself does, s. e.g. qtbase
commit [1]:

    commit 6ec35ff8c56222a317b838b6aac390eb6974de85
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat Jul 8 15:16:12 2023 +0200

        a11y: Report app as parent for top-level item views

        As happens for other widgets
        (s. QAccessibleWidget::parentObject), report the app
        as accessible parent for item views that don't
        have another parent set.

        Otherwise, the accessible tree is broken when
        there's a top-level item view:
        The application has the item view as a child,
        but the child does not have any parent set.

        Extend a QListView autotest accordingly.

        Fixes: QTBUG-115135
        Pick-to: 6.6 6.5
        Change-Id: Ie06874681180a30fc6248dc98f80c4158d837278
        Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>

[1] https://code.qt.io/cgit/qt/qtbase.git/commit/?id=6ec35ff8c56222a317b838b6aac390eb6974de85

Change-Id: Ia20bf1d5311dd918eb7563161745cdb70edc0787
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154982
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>