xmloff: fix import of CharComplexColor - add StylePropertiesContext
Import of CharComplexColor in Calc didn't work because the class
XMLComplexColorContext wasn't used.

This introduces a new class StylePropertiesContext, because the
plain SvXMLPropertySetContext is not enough as we can't define to
use XMLComplexColorContext when XML_CHAR_COMPLEX_COLOR is used on
that class.

Also adds a import/export test for CharComplexColor.

Change-Id: I765ac71a801a5e9bcb105aaea73637df5e601b85
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155008
Tested-by: Tomaž Vajngerl <quikee@gmail.com>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit b543463f639ecee793484e8b419147d565a53d52)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155176
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>