sw: layout: invalidate row after fixed-height row too
In SwLayoutFrame::GrowFrame(), invalidate the position of not only the
next frame, but continue to invalidate until there is a frame with a
non-fixed height.

The idea with invalidating only the next frame is that the next frame
will then be formatted and this will then invalidate the position of the
frame after that.

But with a table row with fixed height, this doesn't work:
SwLayoutFrame::GrowFrame() returns without doing anything, not invalidating
next frame; the next-next row is never repositioned and overlaps the
next row.

(this reproduces as a user-visible problem only when using the API to
 export to PDF, and only in a 6.3-based downstream branch, because
 in SwXTextDocument::getRendererCount() view settings are applied and
 this will differ on the IsParagraph() setting and in master this causes
 all pages to be invalidated and another run of the layout fixes the
 problem; also on the downstream branch backports of commit
 383032c50a3e3354f04200ce984a47ab9d2c5c67 "tdf#123583 use TaskStopwatch
 for Writer Idle loop" and commit
 c605283ad6785dea762feab5fdffd9d27e75c292 "sw: fix spurious layout
 invalidation from ~SwCallLink()" are required to reproduce.)

Change-Id: Ic8a0aa23c496eeab2f3fd87bff7212c8d8ca1cfe
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155017
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>