ScriptForge (SF_Platform) new UserData property
The UserData property lists the content of the
"User Data" oage of the Options dialog.

The content is returned:
- in Basic: in a SF_Dictionary instance
- in Python: in a dict[] instance.

The list of the available keys is:
 city, company, country, email, encryptionkey,
 encrypttoself, fax, firstname, homephone,
 initials, lastname, officephone, position,
 postalcode, signingkey, state, street, title
Many are different from the UNO/priginal values
to make them more user understandable.

This commit will require an update of the SF_Platform
help page.

Both Basic and Python user scripts are supported.

Change-Id: If645579ee9109a1b8180da5a5f3a71979bb5ca59
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155024
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins