partial Revert "tdf#152463 Turn on AutoSave by default"
This reverts 7.6 commit 5cb7fed2a5a02ff1cb4551752a0bd8d3001a1f22,
because this setting is not used in 24.2, and was reverted
for 7.6 as well. So no point in changing a deprecated setting.

In 24.2, the officecfg was corrected to point to the traditional
Recovery::AutoSave::TimeIntervall (which as broken in 7.3,
but had always been a bit of a mess.)

This original setting already has AutoSave enabled by default,
so this patch is NOT turning off AutoSave for 24.2.

Change-Id: I392d5a1511792ec0a42add11a37de8f7e0181631
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155057
Tested-by: Justin Luth <jluth@mail.com>
Reviewed-by: Justin Luth <jluth@mail.com>