tdf#57414 autorecovery: if no recovery doc, remove from RecoveryList
The document has just been saved
(perhaps manually, perhaps automatically)
and so the recovery ODF files are being removed.

So remove from RecoveryList since there is nothing
recoverable at this point.

Sure the document is open, and perhaps
there is even a modification already,
but if there are no recovery files available,
then why display it to the user?

Change-Id: I309cd7df573c960f75c2240c8b5cf171fc1d95be
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155059
Tested-by: Justin Luth <jluth@mail.com>
Reviewed-by: Justin Luth <jluth@mail.com>