tdf#156506 Fix failure to draw copy border on Windows
Revert portion of commit 9d68c794d67259a38de1465090f6f1e7fb588d62
that stops the animated border when a copy action is performed in
a different document by cancelling the animated border.

While the change worked on macOS and Linux, on Windows it stops
the currently active copy border so revert this fix.

Change-Id: I0dbdc8b1bf3077eecadbbe591a27e60ef3d48b7a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155039
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Reviewed-by: Patrick Luby <plubius@neooffice.org>
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>
Tested-by: Thorsten Behrens <thorsten.behrens@allotropia.de>
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>