tdf#156350 sw floattable: fix bad additional <draw:frame> in ODT with layout
The document has a floating table, split on two pages. Saving as ODT
creates two <draw:frame> elements, while we only expect one.

The document model is correct, SwDoc::mpSpzFrameFormatTable only
contains one frame, and in general the ODT export works with the SwDoc,
but CollectFrameAtNode() uses the layout to help performance, and the
layout is available when saving interactively (i.e. not --convert-to),
which visits both layout frames of the same frame format.

Fix the problem by ignoring follow fly frames in
lcl_CollectFrameAtNodeWithLayout(), just working from master should
ensure there is no duplication.

This is similar to 4721729fba32a02683ecc930b630491599f8c6c5
(SwXParaFrameEnumeration: ignore textboxes, 2014-05-27), but that was
for fly frames of draw shapes, and this is for split flys.

Change-Id: I09729c0b0f9afd694e3cbf8886ccbc530bfc9674
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155081
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins