tdf#156372 sw: Stop adding GetLower to every cell in header table
The code was doubling the lower spacing of the last paragraph
in every cell of the table. This change is to not add the space
when in tables, since that was intended to only apply
to the last paragraph in the header.

make CppunitTest_sw_ooxmlexport19 CPPUNIT_TEST_NAME=testTdf156372

Change-Id: I4f676820634390416e6e96ea4149e48fbf0d8908
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155084
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Tested-by: Jenkins