tdf#156260 sw floattable: avoid overlapping flys on anchor change
Going to the end of the document, typing "dt" and F3 collapses the
multi-page floating table into a single page, with overlapping text.

In practice we insert a new paragraph before the "dt" one, insert the
dummy text there and then re-anchor the floating table from the old
paragraph to the new one. Such re-anchoring isn't really meant to be
done with floating tables, which are always just anchored in the next
paragraph in practice. An additional problem is that the amount of fly
frames created depends on the position of the first fly frame, so if the
anchor changes, we may need a different number of split fly frames.

Fix the problem by not trying to reuse the old fly frames on anchor
change: delete the old frames, change the anchor and finally create the
new frames, which leads to correct layout with complicated logic trying
to update the old fly chain to the new anchor.

The original document still puts some dummy text on page 4 instead of
starting it on page 5, that part is still unfixed.

Change-Id: I74549e2ea8ca0d06a9e4814a58aaa8ca476263dc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155122
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins